refactor: Extract backend URI fetching into reusable utility
* Removed redundant fetchConfig function from SensorStatistic component
* Introduced getWebsocketBase utility function for centralized backend URI retrieval
* Updated SensorStatistic to leverage new utility for improved maintainability
* refactor SensorStatistic component to new location
    * Removed deprecated ui.tsx file from /sensors
    * Updated import path in page.tsx to reflect new location of SensorStatistic
